Move syntactic sugar methods to dedicated class with improved Javadoc
