+++ src/main/java/io/github/ricoapon/readableregex/GroupBuilder.java
@@ -59,55 +59,4 @@ public interface GroupBuilder {
      * @return This builder.
      */
     ReadableRegex endGroup();
-
-    /**
-     * Syntactic sugar for {@code .startGroup().add(regexBuilder).endGroup()}.
-     * @param regexBuilder The regular expression.
-     * @return This builder.
-     */
-    ReadableRegex group(ReadableRegex regexBuilder);
-
-    /**
-     * Syntactic sugar for {@code .startGroup(groupName).add(regexBuilder).endGroup()}.
-     * @param groupName    The name of the group.
-     * @param regexBuilder The regular expression.
-     * @return This builder.
-     */
-    ReadableRegex group(String groupName, ReadableRegex regexBuilder);
-
-    /**
-     * Syntactic sugar for {@code .startPositiveLookbehind().add(regexBuilder).endGroup()}.
-     * @param regexBuilder The regular expression.
-     * @return This builder.
-     */
-    default ReadableRegex positiveLookbehind(ReadableRegex regexBuilder) {
-        return startPositiveLookbehind().add(regexBuilder).endGroup();
-    }
-
-    /**
-     * Syntactic sugar for {@code .startNegativeLookbehind().add(regexBuilder).endGroup()}.
-     * @param regexBuilder The regular expression.
-     * @return This builder.
-     */
-    default ReadableRegex negativeLookbehind(ReadableRegex regexBuilder) {
-        return startNegativeLookbehind().add(regexBuilder).endGroup();
-    }
-
-    /**
-     * Syntactic sugar for {@code .startPositiveLookahead().add(regexBuilder).endGroup()}.
-     * @param regexBuilder The regular expression.
-     * @return This builder.
-     */
-    default ReadableRegex positiveLookahead(ReadableRegex regexBuilder) {
-        return startPositiveLookahead().add(regexBuilder).endGroup();
-    }
-
-    /**
-     * Syntactic sugar for {@code .startNegativeLookahead().add(regexBuilder).endGroup()}.
-     * @param regexBuilder The regular expression.
-     * @return This builder.
-     */
-    default ReadableRegex negativeLookahead(ReadableRegex regexBuilder) {
-        return startNegativeLookahead().add(regexBuilder).endGroup();
-    }
 }
